More errors removed

+++ src/main/java/mcjty/rftools/blocks/elevator/ElevatorStopSound.java
@@ -44,6 +44,6 @@ public class ElevatorStopSound extends TickableSound {
             donePlaying = true;
             return;
         }
-        volume = (float) ElevatorConfiguration.baseElevatorVolume.get();
+        volume = (float) (double) ElevatorConfiguration.baseElevatorVolume.get();
     }
 }
